Tags, UI Changes and Bloons theme
Added 4 new tags, Valid, Invalid, Selected. and Outline. Altered the selection code in UI manager to detect these tags. When a player selects a card, it is tagged as selected. Right now all other cards in that player's deck are darkened and marked as invalid, while all cards in other players decks are marked as valid. Invalid cards can not be selected while valid cards can. This can be used to communicate rules and boundaries to the player in an organic way. Also changed music to Bloons TD5 main theme, changed the outline from a copy of the card to its own prefab, moved the GenerateCardSpacing method to UIManager, Fixed a bug in GenerateCardSpacing where adding to the spacing float would shift the cards to the right, mirrored player2's card's sprites, and added logic to outlines to interact with the new tags.