Added Menu System
Added Menu system with semi working buttons. So far only the flip action is fully functional. The switch action is half implemented, and all the other actions need to be implemented fully. At line 1169 in game manager the ExecutePendingAction method receives signals from the menu and uses the highlight system to mark cards in order to perform action. All the method needs is for all the cases to be filled out with the arguments from each action's method.